--- scene/test/Level.tscn
@@ -1,19 +1,0 @@
-[gd_scene load_steps=4 format=3 uid="uid://boubm27mo5xd3"]
-
-[ext_resource type="Script" path="res://script/test/Level.cs" id="1_kpqyr"]
-[ext_resource type="PackedScene" uid="uid://u76umsqvikue" path="res://scene/test/LevelTile.tscn" id="2_5h3jq"]
-
-[sub_resource type="FastNoiseLite" id="FastNoiseLite_47lrw"]
-noise_type = 0
-frequency = 0.07
-fractal_type = 2
-fractal_octaves = 1
-
-[node name="Level" type="Node2D"]
-script = ExtResource("1_kpqyr")
-_levelNoise = SubResource("FastNoiseLite_47lrw")
-_levelTile = ExtResource("2_5h3jq")
-_camera = Object(Camera2D,"_import_path":NodePath(""),"unique_name_in_owner":false,"process_mode":0,"process_priority":0,"process_physics_priority":0,"process_thread_group":0,"editor_description":"","visible":true,"modulate":Color(1, 1, 1, 1),"self_modulate":Color(1, 1, 1, 1),"show_behind_parent":false,"top_level":false,"light_mask":1,"visibility_layer":1,"z_index":0,"z_as_relative":true,"y_sort_enabled":false,"texture_filter":0,"texture_repeat":0,"material":null,"use_parent_material":false,"position":Vector2(0, 0),"rotation":0.0,"scale":Vector2(1, 1),"skew":0.0,"offset":Vector2(0, 0),"anchor_mode":1,"ignore_rotation":true,"enabled":true,"zoom":Vector2(1, 1),"process_callback":1,"limit_left":-10000000,"limit_top":-10000000,"limit_right":10000000,"limit_bottom":10000000,"limit_smoothed":false,"position_smoothing_enabled":false,"position_smoothing_speed":5.0,"rotation_smoothing_enabled":false,"rotation_smoothing_speed":5.0,"drag_horizontal_enabled":false,"drag_vertical_enabled":false,"drag_horizontal_offset":0.0,"drag_vertical_offset":0.0,"drag_left_margin":0.2,"drag_top_margin":0.2,"drag_right_margin":0.2,"drag_bottom_margin":0.2,"editor_draw_screen":true,"editor_draw_limits":false,"editor_draw_drag_margin":false,"script":null)
-
-
-[node name="Camera2D" type="Camera2D" parent="."]
