+++ scene/test/TestLevel.tscn
@@ -1,0 +1,18 @@
+[gd_scene load_steps=4 format=3 uid="uid://boubm27mo5xd3"]
+
+[ext_resource type="Script" path="res://script/test/TestLevel.cs" id="1_kpqyr"]
+[ext_resource type="PackedScene" uid="uid://u76umsqvikue" path="res://scene/test/TestLevelTile.tscn" id="2_5h3jq"]
+
+[sub_resource type="FastNoiseLite" id="FastNoiseLite_47lrw"]
+noise_type = 0
+frequency = 0.07
+fractal_type = 2
+fractal_octaves = 1
+
+[node name="Level" type="Node2D" node_paths=PackedStringArray("_camera")]
+script = ExtResource("1_kpqyr")
+_levelNoise = SubResource("FastNoiseLite_47lrw")
+_levelTile = ExtResource("2_5h3jq")
+_camera = NodePath("Camera2D")
+
+[node name="Camera2D" type="Camera2D" parent="."]
